05_Advanced_RAG/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
python-dotenv>=1.0.0
chromadb>=0.5.0
pypdf>=4.0.0
